+++ app/containers/AccountPage.js
@@ -4,7 +4,6 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { push } from 'react-router-redux';
 import Account from '../components/Account';
-import accountActions from '../redux/account/actions';
 import { links } from '../config';
 import { openLink } from '../lib/platform';
 
@@ -16,14 +15,11 @@ const mapStateToProps = (state: ReduxState) => ({
   accountExpiry: state.account.expiry,
 });
 const mapDispatchToProps = (dispatch: ReduxDispatch, props: SharedRouteProps) => {
-  const { backend } = props;
   const { push: pushHistory } = bindActionCreators({ push }, dispatch);
-  const { logout } = bindActionCreators(accountActions, dispatch);
-
   return {
-    updateAccountExpiry: () => backend.updateAccountExpiry(),
+    updateAccountExpiry: () => props.app.updateAccountExpiry(),
     onLogout: () => {
-      logout(backend);
+      props.app.logout();
     },
     onClose: () => {
       pushHistory('/settings');
