+++ app/containers/PreferencesPage.js
@@ -13,12 +13,11 @@ const mapStateToProps = (state: ReduxState) => ({
 });
 
 const mapDispatchToProps = (dispatch: ReduxDispatch, props: SharedRouteProps) => {
-  const { backend } = props;
   const { push: pushHistory } = bindActionCreators({ push }, dispatch);
   return {
     onClose: () => pushHistory('/settings'),
     onChangeAllowLan: (allowLan) => {
-      backend.setAllowLan(allowLan);
+      props.app.setAllowLan(allowLan);
     },
   };
 };
