+++ app/index.js
@@ -1,7 +1,13 @@
-import React from 'react';
+// @flow
+
 import RX from 'reactxp';
 import App from './app';
 
+const app = new App();
+const view = app.renderView();
+
+app.connect();
+
 RX.App.initialize(true, true);
-RX.UserInterface.setMainView(<App />);
+RX.UserInterface.setMainView(view);
 RX.UserInterface.useCustomScrollbars(true);
